+++ node.gyp
@@ -247,8 +247,8 @@
         'src/aliased_buffer.h',
         'src/async_wrap.h',
         'src/async_wrap-inl.h',
-        'src/base-object.h',
-        'src/base-object-inl.h',
+        'src/base_object.h',
+        'src/base_object-inl.h',
         'src/connection_wrap.h',
         'src/connect_wrap.h',
         'src/env.h',
