+++ src/async_wrap-inl.h
@@ -25,7 +25,7 @@
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
 #include "async_wrap.h"
-#include "base-object-inl.h"
+#include "base_object-inl.h"
 #include "node_internals.h"
 
 namespace node {
