+++ src/async_wrap.h
@@ -24,7 +24,7 @@
 
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
-#include "base-object.h"
+#include "base_object.h"
 #include "v8.h"
 
 #include <stdint.h>
