+++ src/base_object.h
