+++ src/node_contextify.cc
@@ -21,7 +21,7 @@
 
 #include "node_internals.h"
 #include "node_watchdog.h"
-#include "base-object-inl.h"
+#include "base_object-inl.h"
 #include "v8-debug.h"
 
 namespace node {
