+++ src/node_crypto.h
@@ -32,7 +32,7 @@
 
 #include "env.h"
 #include "async_wrap-inl.h"
-#include "base-object-inl.h"
+#include "base_object-inl.h"
 
 #include "v8.h"
